+++ server/controller/authController.ts
@@ -1,8 +1,9 @@
 import { RequestHandler } from "express";
 
 export const signup: RequestHandler = (req, res, next) => {
-    console.log("sign up");
-
+    const sessionData = req.session;
+    console.log(sessionData);
+        
     // sign up
     res.status(202).json({
         user: req.user,
