+++ server/interfaces/userInterface.ts
@@ -1,0 +1,9 @@
+interface UserDoc extends Document {
+    _id: string;
+    email: string;
+    password: string;
+    isModified: (pw: string) => Promise<boolean>;
+    matchPassword: (pw: string) => Promise<boolean>;
+}
+
+export default UserDoc;
