+++ server/package.json
@@ -7,6 +7,8 @@
     "body-parser": "^1.20.2",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "express-session": "^1.17.3",
+    "jsonwebtoken": "^9.0.2",
     "mongoose": "^7.5.2",
     "passport": "^0.6.0",
     "passport-local": "^1.0.0"
@@ -14,6 +16,8 @@
   "devDependencies": {
     "@types/bcrypt": "^5.0.0",
     "@types/express": "^4.17.17",
+    "@types/express-session": "^1.17.7",
+    "@types/jsonwebtoken": "^9.0.3",
     "@types/node": "^20.6.2",
     "@types/passport-local": "^1.0.35",
     "nodemon": "^3.0.1",
